refactor: 프로필 사진 설정 바텀시트 분리
